+++ src/main/scala/com/github/phisgr/gatling/pb/package.scala
@@ -6,10 +6,10 @@ import io.gatling.core.session._
 import scalapb.lenses.Lens._
 import scalapb.lenses.{Lens, Mutation, Updatable}
 
-import scala.collection.GenTraversableOnce
+import scala.collection.SeqOps
+import scala.collection.immutable.SetOps
 
 package object pb {
-
   implicit class EpxrLens[A, B](val l: Lens[A, B]) extends AnyVal {
     def :~(e: Expression[B]): Expression[Mutation[A]] = e.map(l := _)
     def modifyExpr(e: Expression[B => B]): Expression[Mutation[A]] = e.map(l.modify)
@@ -19,33 +19,29 @@ package object pb {
   implicit class EpxrSeqLens[A, B](val l: Lens[A, Seq[B]]) extends AnyVal {
     private def ll = new EpxrSeqLikeLens(l)
     def :+~(e: Expression[B]): Expression[Mutation[A]] = ll :+~ e
-    def :++~(e: Expression[GenTraversableOnce[B]]): Expression[Mutation[A]] = ll :++~ e
+    def :++~(e: Expression[IterableOnce[B]]): Expression[Mutation[A]] = ll :++~ e
     def foreachExpr(f: Lens[B, B] => Expression[Mutation[B]]): Expression[Mutation[A]] = ll.foreachExpr(f)
   }
 
-  implicit class EpxrSeqLikeLens[A, B, Coll[B] <: collection.SeqLike[B, Coll[B]]](
+  implicit class EpxrSeqLikeLens[A, B, Coll[B] <: SeqOps[B, Coll, Coll[B]]](
     val l: Lens[A, Coll[B]]
   ) extends AnyVal {
-    type CBF = collection.generic.CanBuildFrom[Coll[B], B, Coll[B]]
-
-    def :+~(e: Expression[B])(implicit cbf: CBF): Expression[Mutation[A]] =
+    def :+~(e: Expression[B]): Expression[Mutation[A]] =
       e.map(l :+= _)
-    def :++~(e: Expression[GenTraversableOnce[B]])(implicit cbf: CBF): Expression[Mutation[A]] =
+    def :++~(e: Expression[IterableOnce[B]]): Expression[Mutation[A]] =
       e.map(l :++= _)
-    def foreachExpr(f: Lens[B, B] => Expression[Mutation[B]])(implicit cbf: CBF): Expression[Mutation[A]] =
+    def foreachExpr(f: Lens[B, B] => Expression[Mutation[B]]): Expression[Mutation[A]] =
       f(Lens.unit).map(m => l.foreach(_ => m))
   }
 
-  implicit class EpxrSetLens[A, B, Coll[B] <: collection.SetLike[B, Coll[B]] with Set[B]](
+  implicit class EpxrSetLens[A, B, Coll[B] <: SetOps[B, Coll, Coll[B]]](
     val l: Lens[A, Coll[B]]
   ) extends AnyVal {
-    type CBF = collection.generic.CanBuildFrom[Coll[B], B, Coll[B]]
-
     def :+~(e: Expression[B]): Expression[Mutation[A]] =
       e.map(l :+= _)
-    def :++~(e: Expression[GenTraversableOnce[B]]): Expression[Mutation[A]] =
+    def :++~(e: Expression[IterableOnce[B]]): Expression[Mutation[A]] =
       e.map(l :++= _)
-    def foreachExpr(f: Lens[B, B] => Expression[Mutation[B]])(implicit cbf: CBF): Expression[Mutation[A]] =
+    def foreachExpr(f: Lens[B, B] => Expression[Mutation[B]]): Expression[Mutation[A]] =
       f(Lens.unit).map(m => l.foreach(_ => m))
   }
 
@@ -94,5 +90,4 @@ package object pb {
   }
 
   implicit def value2ExprUpdatable[A <: Updatable[A]](e: A): ExprUpdatable[A] = new ExprUpdatable(value2Expression(e))
-
 }